header ve CampaignModellerde de düzenleme

+++ lib/src/model/response/campaign_model.dart
@@ -75,12 +75,12 @@ class CampaignProductResultModel {
     this.campaignDiscountPercent,
   });
 
-  int? newPrice;
+  double? newPrice;
   int? productId;
   int? campaignDiscountPercent;
 
   factory CampaignProductResultModel.fromJson(Map<String, dynamic> json) => CampaignProductResultModel(
-    newPrice: json["new_price"],
+    newPrice: json["new_price"].toDouble(),
     productId: json["product_id"],
     campaignDiscountPercent: json["campaign_discount_percent"],
   );
